+++ AKE/Main/src/main.cpp
@@ -1,11 +1,11 @@
-#include "Main/2DSceneExample.h"
+#include "Main/Example2D.h"
 #include "Main/3DSceneExample.h"
 #include "Main/FluidExample.h"
 #include "Main/FractalExample.h"
 
 int main()
 {
-  Scene2DExamples::CollisionDetectionExample(800, 800, 10);
+  Example2D::CollisionDetection(800, 800, 10);
   // SceneExample3D();
   // FluidExample();
   // FractalExample();
